Simplify bridge token implementation and proxy deployment
This uses a less error-prone strategy than copying the storage layout
between the proxy and the token image in order to initialize the token
storage.

This approach is backwards-compatible with proxy implementation for
already-bridged tokens on chain, whilst simplifying going forward.

Note that this deletes off-chain signing permissioning for token
transfers - these are not necessary in the card pay protocol and is
un-needed added complexity. Along with removing some other
bridged token complexity that just increased surface area

CS-2862
CS-2863